+++ rust/main/lander/src/adapter/chains/radix/conf.rs
@@ -103,9 +103,6 @@ mod tests {
                 transaction_overrides: TransactionOverrides::default(),
                 op_submission_config: OpSubmissionConfig::default(),
                 consider_null_transaction_receipt: false,
-                wallet_urls: None,
-                wallet_solidity_urls: None,
-                energy_multiplier: None,
             }),
         );