+++ typescript/infra/config/environments/mainnet3/warp/configGetters/getEniWarpConfigs.ts
@@ -25,7 +25,7 @@ const owners = {
   polygon: '0x3211A1Fea94cd4000Bd82D7C9E9334E51938De1b',
 } as const;
 
-const WARP_FEE_BPS = 8n;
+const WARP_FEE_BPS = 8;
 
 const usdcTokenAddresses = {
   arbitrum: tokens.arbitrum.USDC,
@@ -262,4 +262,37 @@ export async function getEniUsdtWarpConfig(
   configs.push(['eni', eni]);
 
   return Object.fromEntries(configs);
+}
+
+export async function getEni1PieceWarpConfig(
+  routerConfig: ChainMap<RouterConfigWithoutOwner>,
+): Promise<ChainMap<HypTokenRouterConfig>> {
+  const eni: HypTokenRouterConfig = {
+    ...routerConfig.eni,
+    owner: owners.eni,
+    type: TokenType.synthetic,
+    name: 'OnePiece',
+    symbol: '1Piece',
+    decimals: 18,
+    tokenFee: getFixedRoutingFeeConfig(
+      getWarpFeeOwner('eni'),
+      ['bsc'],
+      WARP_FEE_BPS,
+    ),
+  };
+
+  const bsc: HypTokenRouterConfig = {
+    ...routerConfig.bsc,
+    owner: owners.bsc,
+    type: TokenType.collateral,
+    token: tokens.bsc['1Piece'],
+    decimals: 18,
+    name: 'OnePiece',
+    symbol: '1Piece',
+  };
+
+  return {
+    eni,
+    bsc,
+  };
 }